Refactor activity log to be handed by a new event_server
Refactor activity log to be handed by a new event_server instead of the
system_server
