+++ src/api/system_api.js
@@ -242,179 +242,6 @@ module.exports = {
             }
         },
 
-        export_activity_log: {
-            method: 'GET',
-            params: {
-                type: 'object',
-                required: [],
-                properties: {
-                    event: {
-                        type: 'string',
-                    },
-                    events: {
-                        type: 'array',
-                        items: {
-                            type: 'string'
-                        }
-                    },
-                    till: {
-                        format: 'idate'
-                    },
-                    since: {
-                        format: 'idate'
-                    }
-                }
-            },
-            reply: {
-                type: 'string',
-            },
-            auth: {
-                system: 'admin',
-            }
-        },
-
-        read_activity_log: {
-            method: 'GET',
-            params: {
-                type: 'object',
-                required: [],
-                properties: {
-                    event: {
-                        type: 'string',
-                    },
-                    events: {
-                        type: 'array',
-                        items: {
-                            type: 'string'
-                        }
-                    },
-                    till: {
-                        format: 'idate'
-                    },
-                    since: {
-                        format: 'idate'
-                    },
-                    skip: {
-                        type: 'integer',
-                    },
-                    limit: {
-                        type: 'integer',
-                    },
-                }
-            },
-            reply: {
-                type: 'object',
-                required: ['logs'],
-                properties: {
-                    logs: {
-                        type: 'array',
-                        items: {
-                            type: 'object',
-                            required: ['id', 'time', 'level', 'event'],
-                            properties: {
-                                id: {
-                                    type: 'string',
-                                },
-                                time: {
-                                    format: 'idate'
-                                },
-                                level: {
-                                    type: 'string',
-                                },
-                                event: {
-                                    type: 'string',
-                                },
-                                tier: {
-                                    type: 'object',
-                                    required: ['name'],
-                                    properties: {
-                                        name: {
-                                            type: 'string'
-                                        },
-                                        linkable: {
-                                            type: 'boolean'
-                                        },
-                                    }
-                                },
-                                node: {
-                                    type: 'object',
-                                    required: ['name'],
-                                    properties: {
-                                        name: {
-                                            type: 'string'
-                                        },
-                                        linkable: {
-                                            type: 'boolean'
-                                        },
-                                    }
-                                },
-                                bucket: {
-                                    type: 'object',
-                                    required: ['name'],
-                                    properties: {
-                                        name: {
-                                            type: 'string'
-                                        },
-                                        linkable: {
-                                            type: 'boolean'
-                                        },
-                                    }
-                                },
-                                pool: {
-                                    type: 'object',
-                                    required: ['name'],
-                                    properties: {
-                                        name: {
-                                            type: 'string'
-                                        },
-                                        linkable: {
-                                            type: 'boolean'
-                                        },
-                                    }
-                                },
-                                obj: {
-                                    type: 'object',
-                                    required: ['key'],
-                                    properties: {
-                                        key: {
-                                            type: 'string'
-                                        }
-                                    }
-                                },
-                                account: {
-                                    type: 'object',
-                                    required: ['email'],
-                                    properties: {
-                                        email: {
-                                            type: 'string'
-                                        }
-                                    }
-                                },
-                                actor: {
-                                    type: 'object',
-                                    required: ['email'],
-                                    properties: {
-                                        email: {
-                                            type: 'string'
-                                        }
-                                    }
-                                },
-                                desc: {
-                                    type: 'array',
-                                    items: {
-                                        type: 'string',
-                                    }
-                                },
-                            }
-                        }
-                    },
-                }
-            },
-            auth: {
-                system: 'admin',
-            }
-        },
-
         diagnose_system: {
             method: 'GET',
             reply: {
